+++ dsmr_frontend/tests/webinterface/test_dashboard.py
@@ -16,6 +16,8 @@ from dsmr_stats.models.statistics import DayStatistics
 from dsmr_frontend.models.message import Notification
 from dsmr_datalogger.models.reading import DsmrReading
 import dsmr_consumption.services
+from dsmr_frontend.models.settings import FrontendSettings
+from dsmr_weather.models.reading import TemperatureReading
 
 
 class TestViews(TestCase):
@@ -48,6 +50,11 @@ class TestViews(TestCase):
             reverse('{}:dashboard'.format(self.namespace))
         )
         self.assertEqual(response.status_code, 200)
+        self.assertIn('frontend_settings', response.context)
+        self.assertEqual(
+            response.context['frontend_settings'].dashboard_graph_width,
+            FrontendSettings.get_solo().dashboard_graph_width
+        )
 
     @mock.patch('django.utils.timezone.now')
     def test_dashboard_xhr_header(self, now_mock):
@@ -102,53 +109,143 @@ class TestViews(TestCase):
             self.assertIn('consumption', response.context)
 
     @mock.patch('django.utils.timezone.now')
-    def test_dashboard_xhr_graphs(self, now_mock):
-        now_mock.return_value = timezone.make_aware(timezone.datetime(2015, 11, 15))
+    def test_dashboard_xhr_electricity(self, now_mock):
+        now_mock.return_value = timezone.make_aware(timezone.datetime(2018, 7, 1))
+
+        if self.support_data:
+            ElectricityConsumption.objects.create(
+                read_at=timezone.now(),
+                delivered_1=0,
+                returned_1=0,
+                delivered_2=0,
+                returned_2=0,
+                currently_delivered=1.5,
+                currently_returned=0.1,
+                phase_currently_delivered_l1=0.5,
+                phase_currently_delivered_l2=0.25,
+                phase_currently_delivered_l3=0.75,
+            )
+            ElectricityConsumption.objects.create(
+                read_at=timezone.now() - timezone.timedelta(hours=1),
+                delivered_1=0,
+                returned_1=0,
+                delivered_2=0,
+                returned_2=0,
+                currently_delivered=2.5,
+                currently_returned=0.2,
+                phase_currently_delivered_l1=0.75,
+                phase_currently_delivered_l2=0.5,
+                phase_currently_delivered_l3=1.25,
+            )
+
+        response = self.client.get(
+            reverse('{}:dashboard-xhr-electricity'.format(self.namespace)),
+            data={'delivered': True, 'returned': True, 'phases': True}
+        )
+
+        if not self.support_data:
+            return self.assertEqual(response.status_code, 400, response.content)
+
+        self.assertEqual(response.status_code, 200, response.content)
+
+        json_content = json.loads(response.content.decode("utf8"))
+        self.assertEqual(
+            json_content,
+            {
+                'read_at': ['Sat 23:00', 'Sun 0:00'],
+                'currently_delivered': [2500.0, 1500.0],
+                'currently_returned': [200.0, 100.0],
+                'phases': {
+                    'l1': [750.0, 500.0],
+                    'l2': [500.0, 250.0],
+                    'l3': [1250.0, 750.0],
+                }
+            }
+        )
+
+        # Branch tests for each option.
+        response = self.client.get(
+            reverse('{}:dashboard-xhr-electricity'.format(self.namespace)),
+            data={'delivered': True, 'returned': True, 'phases': False}
+        )
+        self.assertEqual(response.status_code, 200, response.content)
+
+        json_content = json.loads(response.content.decode("utf8"))
+        self.assertNotEqual(json_content['read_at'], [])
+        self.assertNotEqual(json_content['currently_delivered'], [])
+        self.assertNotEqual(json_content['currently_returned'], [])
+        self.assertEqual(json_content['phases']['l1'], [])
+        self.assertEqual(json_content['phases']['l2'], [])
+        self.assertEqual(json_content['phases']['l3'], [])
+
+        response = self.client.get(
+            reverse('{}:dashboard-xhr-electricity'.format(self.namespace)),
+            data={'delivered': False, 'returned': False, 'phases': False}
+        )
+        json_content = json.loads(response.content.decode("utf8"))
+        self.assertNotEqual(json_content['read_at'], [])
+        self.assertEqual(json_content['currently_delivered'], [])
+        self.assertEqual(json_content['currently_returned'], [])
+        self.assertEqual(json_content['phases']['l1'], [])
+        self.assertEqual(json_content['phases']['l2'], [])
+        self.assertEqual(json_content['phases']['l3'], [])
+
+    @mock.patch('django.utils.timezone.now')
+    def test_dashboard_xhr_gas(self, now_mock):
+        now_mock.return_value = timezone.make_aware(timezone.datetime(2018, 7, 1))
+
+        if self.support_data:
+            DataloggerSettings.objects.update(track_phases=True)
+            GasConsumption.objects.create(
+                read_at=timezone.now(),
+                delivered=0,
+                currently_delivered=1,
+            )
+
+            GasConsumption.objects.create(
+                read_at=timezone.now() - timezone.timedelta(hours=1),
+                delivered=0,
+                currently_delivered=0.5,
+            )
+
+        response = self.client.get(
+            reverse('{}:dashboard-xhr-gas'.format(self.namespace))
+        )
+        json_content = json.loads(response.content.decode("utf8"))
+
+        if self.support_data:
+            self.assertEqual(json_content, {'currently_delivered': [0.5, 1.0], 'read_at': ['Sat 23:00', 'Sun 0:00']})
+        else:
+            self.assertEqual(json_content, {'read_at': [], 'currently_delivered': []})
+
+    @mock.patch('django.utils.timezone.now')
+    def test_dashboard_xhr_temperature(self, now_mock):
+        now_mock.return_value = timezone.make_aware(timezone.datetime(2018, 7, 1))
 
         if self.support_data:
             weather_settings = WeatherSettings.get_solo()
             weather_settings.track = True
             weather_settings.save()
 
-            # Test with phases feature switched on as well.
-            DataloggerSettings.get_solo()
-            DataloggerSettings.objects.update(track_phases=True)
+            TemperatureReading.objects.create(
+                read_at=timezone.now(),
+                degrees_celcius=20,
+            )
 
-        # Send seperate offset as well.
-        response = self.client.get(
-            reverse('{}:dashboard-xhr-graphs'.format(self.namespace)),
-            data={'electricity_offset': 24, 'gas_offset': 10}
-        )
-        self.assertEqual(response.status_code, 200, response.content)
-
-        # Send invalid offset.
-        response = self.client.get(
-            reverse('{}:dashboard-xhr-graphs'.format(self.namespace)),
-            data={'electricity_offset': 'abc', 'gas_offset': 'xzy'}
-        )
-        self.assertEqual(response.status_code, 400)
+            TemperatureReading.objects.create(
+                read_at=timezone.now() - timezone.timedelta(hours=1),
+                degrees_celcius=30,
+            )
 
         response = self.client.get(
-            reverse('{}:dashboard-xhr-graphs'.format(self.namespace))
+            reverse('{}:dashboard-xhr-temperature'.format(self.namespace))
         )
-        self.assertEqual(response.status_code, 200, response.content)
         json_content = json.loads(response.content.decode("utf8"))
 
         if self.support_data:
-            self.assertGreater(
-                len(json_content['electricity_x']), 0
-            )
-            self.assertGreater(
-                len(json_content['electricity_y']), 0
-            )
-
-            self.assertIn('phases_l1_y', json_content)
-            self.assertIn('phases_l2_y', json_content)
-            self.assertIn('phases_l3_y', json_content)
-
-        if self.support_gas:
-            self.assertGreater(len(json_content['gas_x']), 0)
-            self.assertGreater(len(json_content['gas_y']), 0)
+            self.assertEqual(json_content, {'degrees_celcius': [30.0, 20.0], 'read_at': ['Sat 23:00', 'Sun 0:00']})
+        else:
+            self.assertEqual(json_content, {'read_at': [], 'degrees_celcius': []})
 
     def test_dashboard_xhr_notification_read(self):
         view_url = reverse('{}:dashboard-xhr-notification-read'.format(self.namespace))