+++ test-all.sh
@@ -1,16 +1,5 @@
 #!/bin/bash
 
-echo ""
-echo "--- Running Pylama for code audit..."
-pylama
-
-# Abort when audit fails.
-if [ $? -ne 0 ]; then
-    echo "[!] Code audit failed"
-    exit;
-fi
-
-echo "OK"
 
 echo ""
 echo "--- Testing with SQLite (4 processes)..."