+++ docs/changelog.rst
@@ -23,7 +23,7 @@ v1.21.0 - 2018-xx-xx
 
 **Tickets resolved in this release:**
 
-- [`#xxxxx <https://github.com/dennissiemensma/dsmr-reader/issues/xxxxx>`_] yyyyy
+- [`#489 <https://github.com/dennissiemensma/dsmr-reader/issues/489>`_] eCharts improved graphs for data zooming/scrolling
 - [`#434 <https://github.com/dennissiemensma/dsmr-reader/issues/434>`_] Omit gas readings all together
 - [`#264 <https://github.com/dennissiemensma/dsmr-reader/issues/264>`_] Check Dropbox API token and display error messages in GUI